refactor: 컴포넌트 분리 및 Shard 페이지 생성
