+++ src/App.jsx
@@ -1,11 +1,13 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Shared from "./pages/Shared";
+import Folder from "./pages/Folder";
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Shared />} />
+        <Route path="/folder" element={<Folder />} />
       </Routes>
     </BrowserRouter>
   );
